# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Proyectos de Pavimento",
        "description": "Documentación de la API para la gestión de proyectos, abscisas, losas y patologías. Creada para la aplicación móvil y web.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pathology-concept/create": {
            "post": {
                "tags": [
                    "Pathologies"
                ],
                "summary": "Insertar concepto técnico de patología",
                "description": "insertar conceptos tecnicos de patologias",
                "operationId": "581014febe948aa3798223d66bdc1802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "url_image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fisura"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Fisura longitudinal en la losa"
                                    },
                                    "url_image": {
                                        "description": "Imagen de referencia",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Concepto creado exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/pathology-concept": {
            "get": {
                "tags": [
                    "Pathologies"
                ],
                "summary": "Mostrar todos los conceptos técnicos de patologías",
                "description": "Show concept pathology",
                "operationId": "aa368856c1cb2543daf3d672d1a13044",
                "responses": {
                    "200": {
                        "description": "Lista de conceptos técnicos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/pathology-concept/search": {
            "get": {
                "tags": [
                    "Pathologies"
                ],
                "summary": "Buscar conceptos técnicos de patologías",
                "description": "Search pathologies",
                "operationId": "deaf78f7679ac2d5ce89203a0ac0a2c5",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Texto de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    },
                    "404": {
                        "description": "No se encontraron resultados"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
        "/api/pathologies/create": {
            "post": {
                "tags": [
                    "Pathologies"
                ],
                "summary": "Crear nueva patología",
                "description": "Create new pathologie",
                "operationId": "f94c4d8fba3e68934c66aa86e2b2742e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slab_id",
                                    "name",
                                    "url_image",
                                    "long_damage",
                                    "type_long_damage",
                                    "width_damage",
                                    "type_width_damage",
                                    "repair_long",
                                    "type_repair_long",
                                    "repair_width",
                                    "type_repair_width",
                                    "type_damage"
                                ],
                                "properties": {
                                    "slab_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fisura"
                                    },
                                    "url_image": {
                                        "description": "Imagen de la patología",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "long_damage": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "type_long_damage": {
                                        "type": "string",
                                        "example": "m"
                                    },
                                    "width_damage": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.3
                                    },
                                    "type_width_damage": {
                                        "type": "string",
                                        "example": "cm"
                                    },
                                    "repair_long": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2
                                    },
                                    "type_repair_long": {
                                        "type": "string",
                                        "example": "m"
                                    },
                                    "repair_width": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 0.2
                                    },
                                    "type_repair_width": {
                                        "type": "string",
                                        "example": "cm"
                                    },
                                    "type_damage": {
                                        "type": "string",
                                        "example": "Fisura"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patología creada exitosamente"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Bearer Token (Laravel Sanctum). Incluir el token en el header 'Authorization'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}